nowruz_app/requirements.txt
flask
langchain_openai
openai
python-dotenv
gunicorn